requirements.txt
gradio
pillow
moviepy
